+++ apps/admin_panel/assets/src/omg-page-each-account/AccountSettingSubPage.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react'
 import AccountSettingPage from '../omg-page-account-setting'
 export default class AccountSettingSubPage extends Component {
-
   render () {
     return <AccountSettingPage divider={false} />
   }
